WIP: refs #218 Buttons and Gallery
