+++ packages/the-wheels/src/styles/partials/atoms/_button.scss
@@ -1,0 +1,105 @@
+.button {
+  --fz: 16;
+  --px: 2.375;
+  --py: 1.3125;
+  display: inline-block;
+  border-radius: calc(var(--br-normal) * 1px);
+  cursor: pointer;
+  width: fit-content;
+  font-weight: 600;
+  color: #e4e4e4;
+  border-style: solid;
+  border-width: 0;
+  background-color: var(--bg-button-main);
+  padding-top: calc(var(--py) * 1em);
+  padding-bottom: calc(var(--py) * 1em);
+  padding-left: calc(var(--px) * 1em);
+  padding-right: calc(var(--px) * 1em);
+  position: relative;
+
+  &:hover {
+    background-color: var(--bg-button-hover);
+  }
+
+  &:active {
+    background-color: var(--text-main);
+    top: 1px;
+    /* height: calc(fit-content - 10px); */
+  }
+  &[disabled] {
+    text-decoration-line: line-through;
+    background: #989898;
+    pointer-events: none;
+    &:after {
+      content: '';
+      display: block;
+      position: absolute;
+      right: -1.2rem;
+      top: -1.2rem;
+      // transform: translateY(-50%);
+      width: 2.4rem;
+      height: 2.4rem;
+      background-image: url("data:image/svg+xml;base64,PHN2ZyB3aWR0aD0iNjUiIGhlaWdodD0iODciIHZpZXdCb3g9IjAgMCA2NSA4NyIgZmlsbD0ibm9uZSIgeG1sbnM9Imh0dHA6Ly93d3cudzMub3JnLzIwMDAvc3ZnIj48cGF0aCBmaWxsLXJ1bGU9ImV2ZW5vZGQiIGNsaXAtcnVsZT0iZXZlbm9kZCIgZD0iTTEwLjgzMzMgMzIuNjI1VjIxLjc1QzEwLjgzMzMgOS43Mzc4IDIwLjUzMzggMCAzMi41IDBDNDQuNDY2IDAgNTQuMTY2NyA5LjczNzggNTQuMTY2NyAyMS43NVYzMi42MjVINjVWODdIMFYzMi42MjVIMTAuODMzM1pNMjEuNjY2NyAyMS43NUMyMS42NjY3IDE1Ljc0MzkgMjYuNTE2OSAxMC44NzUgMzIuNSAxMC44NzVDMzguNDgzMSAxMC44NzUgNDMuMzMzMyAxNS43NDM5IDQzLjMzMzMgMjEuNzVWMzIuNjI1SDIxLjY2NjdWMjEuNzVaTTI3LjA4MzMgNzAuNjg3NVY0OC45Mzc1SDM3LjkxNjdWNzAuNjg3NUgyNy4wODMzWiIgZmlsbD0iY3VycmVudENvbG9yIiAvPjwvc3ZnPg==");
+      background-size: contain;
+      background-repeat: no-repeat;
+      background-origin: center center;
+    }
+  }
+
+  &.full {
+    max-width: 64rem;
+    width: 100%;
+    display: block;
+  }
+
+  &.ghost {
+    background-color: transparent;
+    border: var(--text-caption) 1px solid;
+    color: var(--text-main);
+
+    &:hover {
+      color: var(--text-highlight);
+      border-color: var(--text-highlight);
+      padding-top: calc(var(--py) * 1em - 1px);
+      padding-bottom: calc(var(--py) * 1em - 1px);
+      padding-left: calc(var(--px) * 1em - 1px);
+      padding-right: calc(var(--px) * 1em - 1px);
+      border-width: 2px;
+    }
+
+    &:active {
+      background-color: transparent;
+    }
+
+    &[disabled] {
+      text-decoration-line: line-through;
+      border-color: #989898;
+      color: #989898;
+      pointer-events: none;
+    }
+  }
+
+  &.optional {
+    background-color: transparent;
+    border: 1px solid transparent;
+    color: var(--text-main);
+
+    &:hover {
+      color: var(--text-highlight);
+      background: #d8d8d8;
+      border-color: transparent;
+      padding-top: calc(1.3125em - 1px);
+      padding-bottom: calc(1.3125em - 1px);
+      padding-left: calc(2.375em - 1px);
+      padding-right: calc(2.375em - 1px);
+      border-width: 2px;
+    }
+
+    &[disabled] {
+      text-decoration-line: line-through;
+      // border-color: #989898;
+      color: #989898;
+      pointer-events: none;
+    }
+  }
+}
