+++ packages/the-wheels/src/styles/partials/elements/_gallery.scss
@@ -1,0 +1,19 @@
+.gallery {
+  display: flex;
+  flex-flow: row wrap;
+  gap: 4rem;
+  margin-bottom: 3rem;
+
+  .title {
+    display: inline-block;
+    font-weight: bold;
+    font-size: 1.4rem;
+    margin-bottom: 1.6rem;
+    padding-bottom: 0.5rem;
+    border-bottom: 2px solid var(--text-caption);
+  }
+
+  &.full {
+    flex-flow: column;
+  }
+}
