+++ packages/the-wheels/src/styles/partials/utils/_typography.scss
@@ -1,11 +1,15 @@
-.content-text {
+.content {
   // Basic style, to spacing by gap, cause use flexbox.
-  &, * {
+  &-section {
     display: flex;
     flex-flow: column;
     gap: 3rem;
 
     max-width: 40em;
+
+    &.full {
+      min-width: 100%;
+    }
   }
   
   a:not(.button) {
@@ -14,7 +18,7 @@
 
   h1,h2,h3,h4,h5,h6,
   [class^="heading-"] {
-    margin-top: 3rem;
+    margin-top: 4rem;
     line-height: 1.04;
   }
 
@@ -22,6 +26,7 @@
     --fz: 62;
     font-weight: 800;
     color: var(--text-highlight);
+    margin-bottom: 1.25rem;
   }
 
   h2, .heading-2 {
@@ -48,29 +53,29 @@
   } 
 }
 
-@media (720px < width) {
-  .content-text {
-    &, * {
-      max-width: fit-content;
-    }
-  }
-  .content-section {
-    display: block;
-    column-count: 2;
-    column-rule: solid 1px var(--text-caption);
-    column-gap: 3rem;
-    column-fill: balance;
-    column-width: 36em;
-    max-width: 100dvw;
-    height: auto;
-    max-height: calc(100dvh - 12%);
+// @media (720px < width) {
+//   .content-text {
+//     &, * {
+//       max-width: fit-content;
+//     }
+//   }
+//   .content-section {
+//     display: block;
+//     column-count: 2;
+//     column-rule: solid 1px var(--text-caption);
+//     column-gap: 3rem;
+//     column-fill: balance;
+//     column-width: 36em;
+//     max-width: 100dvw;
+//     height: auto;
+//     max-height: calc(100dvh - 12%);
 
-    p > * {
-      display: inline;
-    }
+//     p > * {
+//       display: inline;
+//     }
 
-    h2, .heading-2 {
-      column-span: all;
-    }
-  }
-}
+//     h2, .heading-2 {
+//       column-span: all;
+//     }
+//   }
+// }
