+++ site/src/components/site/global/subpage-nav.njk
@@ -1,5 +1,5 @@
-<nav id="subpage-navigation">
-  <ul>
+<nav class="subpage-nav">
+  <ul class="nav-item">
     <li><a href="/">トップページ</a></li>
     <li><a href="/about/">仮の下層ページ（About）</a></li>
     <li><a href="/kitchen-sink/">Kitchen sink</a></li>
