+++ site/src/components/site/kitchen-sink/02_radio-group.njk
@@ -1,4 +1,4 @@
-<section class="twls-content-container">
+<section class="content-section">
   <h2 lang="en" id="radio-group">RadioGroup</h2>
   <twls-radio-group>
     <twls-radio-group-item
