+++ site/src/components/site/kitchen-sink/kitchen-sink-side-nav.njk
@@ -1,15 +1,37 @@
-<nav class="twls-sidemenu kitchen-sink-sidemenu">
-  <section class="twls-sidemenu-container">
-    <h2 lang="en" class="twls-sidemenu-heading">Components</h2>
-    <ul class="twls-sidemnu-list">
-      <li class="twls-sidemenu-item"><a href="#button">Button</a></li>
-      <li class="twls-sidemenu-item">
-        <a href="#radio-group">RadioGroup</a>
+<nav class="sidebar">
+  <section class="sidebar-container">
+    <h2 lang="en" class="sidebar-heading">Components</h2>
+    <ul class="sidebar-list">
+      <li class="sidebar-item">
+        <a href="#button">
+          <twls-icon-triangle width="7"></twls-icon-triangle>
+          Button</a>
       </li>
-      <li class="twls-sidemenu-item"><a href="#modal">Modal</a></li>
-      <li class="twls-sidemenu-item"><a href="#accordion">Accordion</a></li>
-      <li class="twls-sidemenu-item"><a href="#message-panel">Message Panel</a></li>
-      <li class="twls-sidemenu-item"><a href="#icons">Icon</a></li>
+      <li class="sidebar-item">
+        <a href="#radio-group">
+          <twls-icon-triangle width="7"></twls-icon-triangle>
+          RadioGroup</a>
+      </li>
+      <li class="sidebar-item">
+        <a href="#modal">
+          <twls-icon-triangle width="7"></twls-icon-triangle>
+          Modal</a>
+      </li>
+      <li class="sidebar-item">
+        <a href="#accordion">
+          <twls-icon-triangle width="7"></twls-icon-triangle>
+          Accordion</a>
+      </li>
+      <li class="sidebar-item">
+        <a href="#message-panel">
+          <twls-icon-triangle width="7"></twls-icon-triangle>
+          Message Panel</a>
+      </li>
+      <li class="sidebar-item">
+        <a href="#icons">
+          <twls-icon-triangle width="7"></twls-icon-triangle>
+          Icon</a>
+      </li>
     </ul>
   </section>
 </nav>
