+++ site/src/pages/about/index.njk
@@ -4,7 +4,7 @@
 {% set caption = "All purpose design system with Japanese language oriented." %}
 
 {% block content %}
-<article class="content-text">
+<article class="content">
   <h1>はじめに</h1>
   <section class="content-section">
     <h2>デザインシステムとは</h2>
