+++ site/src/pages/index.njk
@@ -6,7 +6,7 @@
 
 {% block content %}
     <nav class="twls-container horizontal">
-      <a href="/kitchen-sink/" class="twls-button">Kitchen sink</a>
+      <a href="/kitchen-sink/" class="button">Kitchen sink</a>
     </nav>
     <nav class="index-nav-area">
       {% include 'components/site/home/navigation.njk' %}
