+++ site/src/pages/kitchen-sink/index.njk
@@ -5,7 +5,7 @@
 {% endblock %}
 
 {% block content %}
-  <article>
+  <article class="content">
     <h1 lang="en">Kitchen Sink</h1>
     {% include 'components/site/kitchen-sink/01_buttons.njk' %}
     {% include 'components/site/kitchen-sink/02_radio-group.njk' %}
